requirements.txt
aiogram==2.11.2
aiohttp==3.7.3
aioredis==1.3.1
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.0
cchardet==2.1.7
certifi==2020.12.5
chardet==3.0.4
hiredis==1.1.0
idna==3.1
multidict==5.1.0
pkg-resources==0.0.0
python-dotenv==0.15.0
pytz==2020.5
typing-extensions==3.7.4.3
ujson==4.0.1
uvloop==0.14.0
yarl==1.6.3
